correctly handle sprite file renames
Without this change, if a sprite is generated from, say, 3 files:

* boo.png
* mat.png
* zebra.png

The renaming of one of the files will not bust the sprite map cache. If you did this:

* rename `mat.png` to `doormat.png`, load the sprite map asset, then
* rename `doormat.png` to `amat.png`, load the sprite map asset

The last loading of the asset will not correctly regenerate the sprite map.

To solve this, rather than only taking the MD5 hash of the file contents into account when generating a `cache_key`, we now also create an MD5 hash of the filename.