drop support for Rails 4.2.x
Now that 6.0 has been released, 4.2.x is no longer supported (see rails/rails@1c3c4f0). Note that technically 5.0.x and 5.1.x are also no longer supported by the Rails core team, but we are going to maintain support for those, for now.

Rails 4.2.x was also the reason we needed to run off of Bundler 1.x, so that has been removed.